Private Network File Sharing and Media Buffering
This project is designed to create a server using node, ideally hosted
on a Raspberry PI or a low cost computer. The idea is that the node
server will have access to all your music, films files etc. You can the
stream or download these files through a Web UI.